+++ NEWS.md
@@ -1,5 +1,6 @@
 # orgdata 0.7.5
 - Use either `empty` or `tom` to represent regular expression to replace to since Access makes symbol `""` to be invisible (#285)
+- Change function name from `see_org()` to `see_data()` for viewing data in the data warehouse.
 
 # orgdata 0.7.4
 - Delete raw of similar columns with multiple specifications (#282)